update and fix: fetch static api completely
