+++ nishiki_koi_shop/src/assets/css/Account/payment.css
@@ -1,0 +1,36 @@
+.payment-container {
+    width: 100%;
+    margin: 40px 0;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.payment-content {
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+}
+
+.chooseAddress {
+    width: 30%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+}
+
+.chooseAddress .availableAddress,.chooseAddress .newAddress {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+}
+.form-address{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
