--- nishiki_koi_shop/src/assets/css/payment.css
@@ -1,14 +1,0 @@
-.payment-container{
-    width: 100%;
-    margin:40px 0;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-.payment-content{
-    width: 80%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
