+++ nishiki_koi_shop/src/pages/Account/shoppingCart.js
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from "react";
 import {ToastContainer, toast} from "react-toastify";
 import 'react-toastify/dist/ReactToastify.css';
-import '../../assets/css/cart.css'
+import '../../assets/css/Account/cart.css'
 import {Link, useNavigate} from "react-router-dom";
 
 const Cart = () => {
